modified Jukebox.java and State.java

+++ src/main/java/com/musicbee/controllers/PlaylistSceneController.java
@@ -175,22 +175,20 @@ public class PlaylistSceneController implements Initializable {
         if (event.getButton() == MouseButton.PRIMARY) {
             try {
                 if (!row.isEmpty() && row.getItem() != null) {
-                    int index = row.getIndex();
-                    State.setCurrentSongIndex(index);
-                    Song song = State.getSongsInTable().get(index);
+                    Song song = tableList.get(row.getIndex());
 
-                    MediaPlayerControl.prepare(song);
-                    MediaPlayerControl.play();
+                    Jukebox.setCurrentSongIndex(row.getIndex());
+                    Jukebox.setNowPlaying(song);
 
-                    controlPanel.setPause();
-                    State.setCurrentSongName(State.getSongsInTable().get(State.getCurrentSongIndex()).getName());
-                    State.setCurrentSongArtist(State.getSongsInTable().get(State.getCurrentSongIndex()).getArtistName());
-                    updateNames(State.getCurrentSongName(),
-                            State.getCurrentSongArtist());
-                    MediaPlayerControl.getMediaPlayer().setVolume(State.getVolume());
-                    State.setLastSongID(State.getSongsInTable().get(index).getID());
+                    Jukebox.setSong(song);
+                    Jukebox.prepare();
+                    Jukebox.play();
+
+                    State.setLastSongID(song.getID());
+                    Jukebox.setCurrentList(tableList);
+
+                    controlPanel.update(song.getName(), song.getArtistName());
                 }
-                controlPanel.setTimeSlider();
             } catch (Exception e) {
                 System.out.println(e.getMessage());
                 System.out.println(getClass().getName() + ": " + getClass().getEnclosingMethod());
@@ -200,7 +198,9 @@ public class PlaylistSceneController implements Initializable {
             System.out.println(contextMenu.getItems().get(0).getId());
             contextMenu.setOnAction(e -> {
                 try {
-                    Database.deleteSongFromPlaylist(row.getItem().getID(), State.getCurrentPlaylistID());
+                    if (!row.isEmpty() && row.getItem() != null) {
+                        Database.deleteSongFromPlaylist(row.getItem().getID(), State.getCurrentPlaylistID());
+                    }
                 } catch (SQLException ex) {
                     throw new RuntimeException(ex);
                 }
@@ -216,7 +216,6 @@ public class PlaylistSceneController implements Initializable {
         allSongs.addAll(list);
         filteredSongs.addAll(list);
         tableList.addAll(list);
-        State.setSongsInTable(list);
     }
 
     @FXML
@@ -235,7 +234,6 @@ public class PlaylistSceneController implements Initializable {
             }
             tableList.clear();
             tableList.addAll(filteredSongs);
-            State.setSongsInTable(filteredSongs);
         }
     }
 
@@ -249,10 +247,10 @@ public class PlaylistSceneController implements Initializable {
     }
 
     public void onClickLogOut(ActionEvent event) throws IOException, SQLException {
-        MediaPlayer player = MediaPlayerControl.getMediaPlayer();
+        MediaPlayer player = Jukebox.getMediaPlayer();
 
         if (player != null) {
-            MediaPlayerControl.clear();
+            Jukebox.clear();
         }
 
         Database.savePlaybackPosition();
@@ -263,10 +261,5 @@ public class PlaylistSceneController implements Initializable {
 
         SceneSwitcher sceneSwitcher = new SceneSwitcher(FilePaths.SIGN_IN, FilePaths.STYLESHEET);
         sceneSwitcher.switchNow(stage);
-    }
-
-    private void updateNames(String songName, String artist) {
-        controlPanel.getSongName().setText(songName);
-        controlPanel.getArtistName().setText(artist);
     }
 }
